backend/requirements.txt
google-genai
python-dotenv
Pillow
psycopg2-binary
fastapi
uvicorn
python-multipart
